+++ nodalia-cards.manifest.js
@@ -1,8 +1,10 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.9",
-  "contentSha256_12": "0bddf097a6be",
-  "file": "nodalia-cards.bundle.js"
+  "pkgVersion": "1.1.0-alpha.10",
+  "contentSha256_12": "cb8bca9a3e84",
+  "file": "nodalia-cards.bundle.js",
+  "loaderFile": "nodalia-cards.js",
+  "hacsFile": "nodalia-cards-1.1.0-alpha.10.js"
 };
-export const pkgVersion = "1.1.0-alpha.9";
-export const contentSha256_12 = "0bddf097a6be";
+export const pkgVersion = "1.1.0-alpha.10";
+export const contentSha256_12 = "cb8bca9a3e84";
 export const file = "nodalia-cards.bundle.js";